refactor(relay): パイプライン結合テストをtestsディレクトリモジュールに分割
クレート内結合テストの置き場所として relay/tests.rs + relay/tests/ を切り、
ハーネス（harness.rs）とシナリオ（pipeline.rs）を分離。今後の結合テストは
ここにシナリオファイルを追加してハーネスを共有する。
