+++ relay/src/modules/relay/tests/harness.rs
@@ -1,13 +1,10 @@
-//! In-process integration tests for the live forwarding pipeline:
-//! ingress `StreamReader` -> `TrackCache` -> `EgressRunner` (scheduler +
-//! group sender) -> downstream `DataSender`, with no QUIC involved.
-//!
-//! These target the cascading-relay E2E flake "stream ended before receiving
-//! 50 objects": a publisher that answers SUBSCRIBE with a burst of objects on
-//! one subgroup stream and closes it immediately (FIN) races the relay's
-//! subscribe handling, and the downstream subscriber observes a truncated
-//! stream. The tests pin the pipeline behavior at each stage of that race so
-//! a failure names the stage instead of an opaque E2E timeout.
+//! Test harness for the live forwarding pipeline: wires the real ingress
+//! `StreamReader`, `TrackCache`, notifier and `EgressRunner` together with no
+//! QUIC involved. Only the two boundaries are substituted: the upstream
+//! `StreamReceiver` is fed through a channel so tests control exactly when
+//! each object (and the FIN) becomes visible to the ingress, and the
+//! downstream `Publisher` captures every egressed object, tagged with the
+//! downstream uni-stream it was sent on.
 
 use std::sync::{
     Arc,
@@ -37,12 +34,11 @@ use crate::modules::{
     types::TrackKey,
 };
 
-const OBJECT_COUNT: usize = 50;
+pub(super) const OBJECT_COUNT: usize = 50;
 const RECV_TIMEOUT: Duration = Duration::from_secs(3);
 
 // ---------------------------------------------------------------------------
-// Upstream side: a StreamReceiver fed through a channel, so tests control
-// exactly when each object (and the FIN) becomes visible to the ingress.
+// Upstream side
 // ---------------------------------------------------------------------------
 
 type FeedItem = Result<Option<DataObject>, moqt::StreamReceiveError>;
@@ -62,25 +58,26 @@ impl StreamReceiver for ChannelStreamReceiver {
     }
 }
 
-struct UpstreamFeed {
+/// Plays the publisher side of one upstream subgroup stream.
+pub(super) struct UpstreamFeed {
     sender: mpsc::UnboundedSender<FeedItem>,
 }
 
 impl UpstreamFeed {
-    fn header(&self, group_id: u64) {
+    pub(super) fn header(&self, group_id: u64) {
         self.sender
             .send(Ok(Some(make_header(group_id))))
             .expect("ingress feed should be open");
     }
 
-    fn object(&self, index: usize) {
+    pub(super) fn object(&self, index: usize) {
         let delta = if index == 0 { 0 } else { 1 };
         self.sender
             .send(Ok(Some(make_payload_object(delta, ordered_payload(index)))))
             .expect("ingress feed should be open");
     }
 
-    fn fin(&self) {
+    pub(super) fn fin(&self) {
         self.sender
             .send(Ok(None))
             .expect("ingress feed should be open");
@@ -88,12 +85,11 @@ impl UpstreamFeed {
 }
 
 // ---------------------------------------------------------------------------
-// Downstream side: a Publisher whose stream senders capture every egressed
-// object, tagged with the downstream uni-stream they were sent on.
+// Downstream side
 // ---------------------------------------------------------------------------
 
 #[derive(Debug)]
-enum EgressEvent {
+pub(super) enum EgressEvent {
     Object { stream: usize, object: DataObject },
     Closed { stream: usize },
 }
@@ -188,7 +184,7 @@ impl Publisher for CapturingPublisher {
 // Object builders (payloads carry their index so ordering is verifiable).
 // ---------------------------------------------------------------------------
 
-fn ordered_payload(index: usize) -> Bytes {
+pub(super) fn ordered_payload(index: usize) -> Bytes {
     Bytes::from(format!("ordered-object-{index}"))
 }
 
@@ -233,20 +229,20 @@ fn make_downstream_subscription() -> DownstreamSubscription {
 }
 
 // ---------------------------------------------------------------------------
-// Harness wiring the real ingress reader, cache, notifier and egress runner.
+// Pipeline wiring
 // ---------------------------------------------------------------------------
 
-struct PipelineHarness {
-    track_key: TrackKey,
+pub(super) struct PipelineHarness {
+    pub(super) track_key: TrackKey,
+    pub(super) notify_map: Arc<ObjectNotifyProducerMap>,
     cache_store: Arc<TrackCacheStore>,
-    notify_map: Arc<ObjectNotifyProducerMap>,
     opened_sender: mpsc::Sender<StreamOpened>,
     _stream_reader: StreamReader,
     _stop_sender: watch::Sender<bool>,
     stop_receiver: watch::Receiver<bool>,
 }
 
-struct RunningEgress {
+pub(super) struct RunningEgress {
     events: mpsc::UnboundedReceiver<EgressEvent>,
     handle: tokio::task::JoinHandle<()>,
 }
@@ -258,7 +254,7 @@ impl Drop for RunningEgress {
 }
 
 impl PipelineHarness {
-    fn new() -> Self {
+    pub(super) fn new() -> Self {
         let track_key = TrackKey::new("ns", "track");
         let cache_store = Arc::new(TrackCacheStore::new());
         let notify_map = Arc::new(ObjectNotifyProducerMap::new());
@@ -268,8 +264,8 @@ impl PipelineHarness {
         let (stop_sender, stop_receiver) = watch::channel(false);
         Self {
             track_key,
-            cache_store,
             notify_map,
+            cache_store,
             opened_sender,
             _stream_reader: stream_reader,
             _stop_sender: stop_sender,
@@ -279,7 +275,7 @@ impl PipelineHarness {
 
     /// Opens one upstream subgroup stream in the ingress; the returned feed
     /// plays the publisher side of that stream.
-    async fn open_upstream_stream(&self) -> UpstreamFeed {
+    pub(super) async fn open_upstream_stream(&self) -> UpstreamFeed {
         let (feed_sender, feed_receiver) = mpsc::unbounded_channel();
         self.opened_sender
             .send(StreamOpened {
@@ -299,7 +295,10 @@ impl PipelineHarness {
 
     /// Starts the egress runner the way `sequences::subscribe` does and waits
     /// for its readiness signal (production sends SUBSCRIBE_OK only after it).
-    async fn start_egress(&self, largest_location: Option<moqt::Location>) -> RunningEgress {
+    pub(super) async fn start_egress(
+        &self,
+        largest_location: Option<moqt::Location>,
+    ) -> RunningEgress {
         let (event_sender, event_receiver) = mpsc::unbounded_channel();
         let publisher = CapturingPublisher {
             next_stream: Arc::new(AtomicUsize::new(0)),
@@ -331,7 +330,10 @@ impl PipelineHarness {
 
     /// Waits until the ingress has fully ingested a subgroup (its EndOfGroup
     /// notification fired). Subscribe to events before feeding the FIN.
-    async fn wait_end_of_group(&self, events: &mut tokio::sync::broadcast::Receiver<TrackEvent>) {
+    pub(super) async fn wait_end_of_group(
+        &self,
+        events: &mut tokio::sync::broadcast::Receiver<TrackEvent>,
+    ) {
         loop {
             let event = tokio::time::timeout(RECV_TIMEOUT, events.recv())
                 .await
@@ -345,7 +347,7 @@ impl PipelineHarness {
 
     /// Polls the cache until its Largest Location reaches `expected`,
     /// mirroring the moment `sequences::subscribe` resolves it.
-    async fn wait_largest_location(&self, expected: moqt::Location) -> moqt::Location {
+    pub(super) async fn wait_largest_location(&self, expected: moqt::Location) -> moqt::Location {
         let cache = self.cache_store.get_or_create(&self.track_key);
         let deadline = tokio::time::Instant::now() + RECV_TIMEOUT;
         loop {
@@ -365,10 +367,14 @@ impl PipelineHarness {
     }
 }
 
+// ---------------------------------------------------------------------------
+// Collection and assertion helpers
+// ---------------------------------------------------------------------------
+
 /// Receives egress events until `closed_streams` downstream streams closed.
 /// Panics with the events collected so far when the pipeline stalls, so a
 /// lost-object bug reports as "stalled after N objects" rather than a hang.
-async fn collect_until_closed(
+pub(super) async fn collect_until_closed(
     egress: &mut RunningEgress,
     closed_streams: usize,
 ) -> Vec<EgressEvent> {
@@ -391,7 +397,7 @@ async fn collect_until_closed(
 }
 
 /// Payloads sent on one downstream stream, in send order.
-fn payloads_on_stream(events: &[EgressEvent], stream: usize) -> Vec<Bytes> {
+pub(super) fn payloads_on_stream(events: &[EgressEvent], stream: usize) -> Vec<Bytes> {
     events
         .iter()
         .filter_map(|event| match event {
@@ -407,13 +413,13 @@ fn payloads_on_stream(events: &[EgressEvent], stream: usize) -> Vec<Bytes> {
         .collect()
 }
 
-fn stream_closed(events: &[EgressEvent], stream: usize) -> bool {
+pub(super) fn stream_closed(events: &[EgressEvent], stream: usize) -> bool {
     events
         .iter()
         .any(|event| matches!(event, EgressEvent::Closed { stream: s } if *s == stream))
 }
 
-fn header_group_on_stream(events: &[EgressEvent], stream: usize) -> Option<u64> {
+pub(super) fn header_group_on_stream(events: &[EgressEvent], stream: usize) -> Option<u64> {
     events.iter().find_map(|event| match event {
         EgressEvent::Object {
             stream: s,
@@ -423,7 +429,7 @@ fn header_group_on_stream(events: &[EgressEvent], stream: usize) -> Option<u64> 
     })
 }
 
-fn assert_full_ordered_delivery(events: &[EgressEvent]) {
+pub(super) fn assert_full_ordered_delivery(events: &[EgressEvent]) {
     assert_eq!(
         header_group_on_stream(events, 0),
         Some(0),
@@ -441,190 +447,5 @@ fn assert_full_ordered_delivery(events: &[EgressEvent]) {
     assert!(
         stream_closed(events, 0),
         "the downstream stream should be closed after the last object"
-    );
-}
-
-// ---------------------------------------------------------------------------
-// Scenarios
-// ---------------------------------------------------------------------------
-
-/// Baseline: egress is running (largest resolved as None before any object
-/// existed) when the publisher bursts one subgroup of 50 objects and FINs
-/// immediately. Every object must reach the downstream before its stream
-/// closes.
-#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-async fn burst_publish_with_immediate_fin_delivers_all_objects() {
-    for round in 0..100 {
-        let harness = PipelineHarness::new();
-        let mut egress = harness.start_egress(None).await;
-
-        let feed = harness.open_upstream_stream().await;
-        feed.header(0);
-        for index in 0..OBJECT_COUNT {
-            feed.object(index);
-        }
-        feed.fin();
-
-        let events = collect_until_closed(&mut egress, 1).await;
-        let payloads = payloads_on_stream(&events, 0);
-        assert_eq!(
-            payloads.len(),
-            OBJECT_COUNT,
-            "round {round}: downstream stream ended before receiving {OBJECT_COUNT} objects"
-        );
-        assert_full_ordered_delivery(&events);
-    }
-}
-
-/// The egress starts while the burst is already in flight (subscribe-time
-/// largest was resolved as None before the burst reached the relay). The
-/// scheduler must catch up from the cache regardless of whether the
-/// StreamOpened notification fired before it subscribed to track events.
-#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-async fn egress_start_racing_ingest_burst_delivers_all_objects() {
-    for round in 0..100 {
-        let harness = PipelineHarness::new();
-
-        let feed = harness.open_upstream_stream().await;
-        feed.header(0);
-        for index in 0..OBJECT_COUNT {
-            feed.object(index);
-        }
-        feed.fin();
-
-        // No await between feeding and starting the egress: the ingest and
-        // the egress startup interleave differently on every round.
-        let mut egress = harness.start_egress(None).await;
-
-        let events = collect_until_closed(&mut egress, 1).await;
-        let payloads = payloads_on_stream(&events, 0);
-        assert_eq!(
-            payloads.len(),
-            OBJECT_COUNT,
-            "round {round}: downstream stream ended before receiving {OBJECT_COUNT} objects"
-        );
-        assert_full_ordered_delivery(&events);
-    }
-}
-
-/// The subscriber arrives after the whole subgroup was ingested and closed
-/// (fastest possible FIN). A closed subgroup must still be drained in full;
-/// closing must never truncate objects already in the cache.
-#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-async fn closed_subgroup_is_drained_in_full() {
-    let harness = PipelineHarness::new();
-    let mut track_events = harness
-        .notify_map
-        .get_or_create(&harness.track_key)
-        .subscribe();
-
-    let feed = harness.open_upstream_stream().await;
-    feed.header(0);
-    for index in 0..OBJECT_COUNT {
-        feed.object(index);
-    }
-    feed.fin();
-    harness.wait_end_of_group(&mut track_events).await;
-
-    let mut egress = harness.start_egress(None).await;
-
-    let events = collect_until_closed(&mut egress, 1).await;
-    assert_full_ordered_delivery(&events);
-}
-
-/// Reproduces the cascading-relay E2E flake "stream ended before receiving 50
-/// objects" (ordered objects scenario, e.g. CI runs 29805931921, 29481445955,
-/// 29477073181).
-///
-/// Production ordering in `sequences::subscribe`: the upstream ingress starts
-/// on SUBSCRIBE_OK, and only afterwards `accept_downstream_subscription`
-/// resolves the subscribe-time Largest Location from the live cache
-/// (`resolve_subscribe_largest`). A publisher that answers SUBSCRIBE by
-/// bursting objects immediately (the E2E publisher sends all 50 within ~1ms)
-/// lands part of the burst in the cache inside that window. The resolved
-/// largest then points into the burst, the LargestObject filter starts
-/// delivery just after it, and the downstream subscriber - whose SUBSCRIBE is
-/// what triggered this publish, and which observed no content at subscribe
-/// time - permanently loses the head of the group.
-///
-/// The relay must not let objects that arrived after the triggering SUBSCRIBE
-/// shift that subscriber's delivery start.
-#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-#[ignore = "known bug: in-flight burst shifts the resolved largest location and the head of the group is lost; un-ignore with the fix"]
-async fn largest_location_resolved_mid_burst_must_not_skip_head_objects() {
-    const IN_FLIGHT_BEFORE_RESOLVE: usize = 10;
-
-    let harness = PipelineHarness::new();
-
-    // The burst races ahead: the first objects reach the cache before the
-    // subscribe sequence resolves the largest location.
-    let feed = harness.open_upstream_stream().await;
-    feed.header(0);
-    for index in 0..IN_FLIGHT_BEFORE_RESOLVE {
-        feed.object(index);
-    }
-    let largest = harness
-        .wait_largest_location(moqt::Location {
-            group_id: 0,
-            object_id: (IN_FLIGHT_BEFORE_RESOLVE - 1) as u64,
-        })
-        .await;
-
-    // What `accept_downstream_subscription` does now: largest from the live
-    // cache -> EgressStartRequest -> SUBSCRIBE_OK.
-    let mut egress = harness.start_egress(Some(largest)).await;
-
-    for index in IN_FLIGHT_BEFORE_RESOLVE..OBJECT_COUNT {
-        feed.object(index);
-    }
-    feed.fin();
-
-    let events = collect_until_closed(&mut egress, 1).await;
-    assert_full_ordered_delivery(&events);
-}
-
-/// Multi-group variant closer to the media E2E shape: several subgroup
-/// streams in sequence, each closed with FIN right after its last object.
-/// Every group must arrive complete on its own downstream stream.
-#[tokio::test(flavor = "multi_thread", worker_threads = 4)]
-async fn sequential_groups_with_fast_fin_each_arrive_complete() {
-    const GROUPS: u64 = 10;
-    const OBJECTS_PER_GROUP: usize = 5;
-
-    let harness = PipelineHarness::new();
-    let mut egress = harness.start_egress(None).await;
-
-    for group_id in 0..GROUPS {
-        let feed = harness.open_upstream_stream().await;
-        feed.header(group_id);
-        for index in 0..OBJECTS_PER_GROUP {
-            feed.object(index);
-        }
-        feed.fin();
-    }
-
-    let events = collect_until_closed(&mut egress, GROUPS as usize).await;
-
-    let mut delivered_groups = Vec::new();
-    for stream in 0..GROUPS as usize {
-        let group_id = header_group_on_stream(&events, stream)
-            .unwrap_or_else(|| panic!("stream {stream} should carry a subgroup header"));
-        let payloads = payloads_on_stream(&events, stream);
-        let expected: Vec<Bytes> = (0..OBJECTS_PER_GROUP).map(ordered_payload).collect();
-        assert_eq!(
-            payloads, expected,
-            "group {group_id} on stream {stream} must arrive complete and in order"
-        );
-        assert!(
-            stream_closed(&events, stream),
-            "stream {stream} should be closed after its last object"
-        );
-        delivered_groups.push(group_id);
-    }
-    delivered_groups.sort_unstable();
-    assert_eq!(
-        delivered_groups,
-        (0..GROUPS).collect::<Vec<_>>(),
-        "every group must be delivered exactly once"
     );
 }
